requirements.txt
absl-py==0.1.10
apptools==4.4.0
chainer==3.3.0
configobj==5.0.6
configparser==3.5.0
cupy==2.3.0
Cython==0.27.3
decorator==4.2.1
deepdish==0.3.6
dominate==2.3.1
h5py==2.7.1
html5lib==1.0.1
idna==2.6
imageio==2.2.0
ipdb==0.10.3
ipykernel==4.8.0
ipython==4.0.0
ipython-genutils==0.2.0
ipywidgets==7.1.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.2
meshzoo==0.3.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.0
numexpr==2.6.4
numpy==1.14.0
opencv-python==3.4.0.12
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkg-resources==0.0.0
progressbar==2.3
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pybind11==2.2.1
pyface==5.1.0
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
requests==2.18.4
scandir==1.6
scikit-image==0.13.1
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
tables==3.4.2
terminado==0.8.1
testpath==0.3.1
torch==0.3.1
torchfile==0.1.0
torchvision==0.2.0
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
traits==4.6.0
traitsui==5.1.0
urllib3==1.22
visdom==0.1.7
vtk==8.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
